pyproject.toml
[project]
name = "vapi"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
    "dspy>=2.5.0",
    "optuna>=3.6.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
]
